+++ packages/components/src/index.ts
@@ -1,1 +1,2 @@
 export * from './button/lm-prototype-button.js';
+export * from './dropdown/lm-prototype-dropdown.js';